pointerevents/pointerevent_setPointerCapture_invalid_pointerId.html

<!doctype html>
<html>
    <head>
        <title>Pointer Event: gotPiontercapture is fired first.</title>
        <meta content="text/html; charset=UTF-8" http-equiv="Content-Type"/>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/"/>
        <link rel="help" href="http://www.w3.org/wiki/PointerEvents/TestAssertions">

        <meta name="assert" content="When the setPointerCapture method is invoked, if the provided pointerId value does not match any of the active pointers, a DOMException with the name InvalidPointerId must be thrown."/>

        <!-- http://w3c-test.org/resources/testharness.js -->
	    <script src="http://w3c-test.org/resources/testharness.js"></script>
		<!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>
        <script type="text/javascript">
            setup({ explicit_done: true });

            var detected_pointertypes = {};
            add_completion_callback(showPointerTypes);


            var INVALID_POINTERID = -39548;
            var target0;
            var listener;
            var test_pointerCapture_invalidPointerId;

            function run () {

                target0 = document.getElementById("target0");
                target0.style["touchAction"] = "none";
                listener = document.getElementById("complete-notice");

                test_pointerCapture_invalidPointerId = async_test("setPointerCapture: DOMException InvalidPointerId");  // set up test harness

                // add listener to pointerdown
                on_event(target0, pointerPrefix["pointerdown"], function (event) {
                    check_PointerEvent(event);
                    detected_pointertypes[event.pointerType] = true;

                    
                    try {
                        listener.setPointerCapture(INVALID_POINTERID);

                        test_pointerCapture_invalidPointerId.step(function () {
                            assert_true(false,  "DOMException: InvalidPointerId should have been thrown.");
                        });
                    } catch (e) {
                        test_pointerCapture_invalidPointerId.step(function () {
                            assert_equals(e.message, "InvalidPointerId" , e.message + " is the correct DOMException");
                        });
                    }

                    test_pointerCapture_invalidPointerId.done(); // complete test
                });
            }

            function showPointerTypes() {
                var complete_notice = document.getElementById("complete-notice");
                var pointertype_log = document.getElementById("pointertype-log");
                var pointertypes = Object.keys(detected_pointertypes); pointertype_log.innerHTML = pointertypes.length ? pointertypes.join(",") : "(none)";

                if (test_pointerCapture_invalidPointerId.status == test_pointerCapture_invalidPointerId.NOTRUN
                     || test_pointerCapture_invalidPointerId.status == test_pointerCapture_invalidPointerId.TIMEOUT) {
                    complete_notice.innerHTML = "Test FAIL due to time out.  This is due to the browser not supporting pointer, or the manual steps were not run to completion.";
                    complete_notice.style.background = "red";
                }
                complete_notice.style.display = "block";
            }
        </script>

        <style>
            #target0 {
                margin: 0em;
                padding: 2em;
                background: yellow;
                border: 1px solid orange;
            }

            #complete-notice {
                border: 1px solid #0a0;
                display: none;
            }

            #pointertype-log {
                font-weight: bold;
            }
        </style>
    </head>
    <body onload="run()">
        <h1> Pointer Event: DOMException InvalidPointerId</h1>
        <h4> 
            Test Description: 
            When the setPointerCapture method is invoked, if the provided pointerId value does not match any of the active pointers, a DOMException with the name InvalidPointerId must be thrown. 
        </h4>
        <br />
        <div id="target0">
            Use the mouse, touch or pen to contact this box.
        </div>
        <div id="complete-notice">
            <p>Test complete:  Scroll to Summary to view Pass/Fail Results.</p>
            <p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
            <p>Would you like to <a href="#">run the tests again</a> with a different pointer type?</p>
        </div>
        <div id="log"></div>
    </body>
</html>
